requirements.txt
python-owasp-zap-v2.4==0.0.20
requests>=2.32.0
beautifulsoup4>=4.12.3
selenium>=4.23.1
pandas>=2.2.2
scikit-learn>=1.5.1
joblib>=1.4.2
reportlab>=4.2.2
jinja2>=3.1.4
python-dotenv>=1.0.1
tabulate>=0.9.0

